requirements.txt
nltk
numpy
couchdb
facebook-sdk
